requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
groq==0.4.2
pinecone-client==3.0.2
sentence-transformers==2.5.1
